Рефакторинг проекта (удаление лишних файлов и перемещение некоторых). Исправление waring.
