+++ .gitignore
@@ -2,3 +2,5 @@
 /robots/.settings
 /robots/bin
 eclipse.bat
+/target
+/.idea
