--- .idea/Game_OOP.iml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module type="JAVA_MODULE" version="4">
-  <component name="AdditionalModuleElements">
-    <content url="file://$MODULE_DIR$" dumb="true">
-      <sourceFolder url="file://$MODULE_DIR$/robots/src" isTestSource="false" />
-      <sourceFolder url="file://$MODULE_DIR$/robots/tests" isTestSource="true" />
-    </content>
-  </component>
-</module>
