+++ src/gui/GameWindow.java
@@ -1,38 +1,24 @@
 package gui;
 
 import java.awt.BorderLayout;
-import java.awt.EventQueue;
-import java.awt.TextArea;
 import javax.swing.JInternalFrame;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.event.InternalFrameEvent;
 import javax.swing.event.InternalFrameListener;
-import log.LogChangeListener;
-import log.LogEntry;
-import log.LogWindowSource;
 
-public class LogWindow extends JInternalFrame implements LogChangeListener {
-    private LogWindowSource m_logSource;
-    private TextArea m_logContent;
+public class GameWindow extends JInternalFrame {
 
-    public LogWindow(LogWindowSource logSource) {
-        super("Протокол работы", true, true, true, true);
-        m_logSource = logSource;
-        m_logSource.registerListener(this);
-        m_logContent = new TextArea("");
-        m_logContent.setSize(200, 500);
-
+    public GameWindow() {
+        super("Игровое поле", true, true, true, true);
+        GameVisualizer m_visualizer = new GameVisualizer();
         JPanel panel = new JPanel(new BorderLayout());
-        panel.add(m_logContent, BorderLayout.CENTER);
+        panel.add(m_visualizer, BorderLayout.CENTER);
         getContentPane().add(panel);
         pack();
-        updateLogContent();
 
-        // Устанавливаем поведение при закрытии
         setDefaultCloseOperation(JInternalFrame.DO_NOTHING_ON_CLOSE);
 
-        // Добавляем слушатель для подтверждения закрытия
         addInternalFrameListener(new InternalFrameListener() {
             @Override
             public void internalFrameClosing(InternalFrameEvent e) {
@@ -68,19 +54,5 @@ public class LogWindow extends JInternalFrame implements LogChangeListener {
             dispose(); // Закрываем окно, если пользователь подтвердил
         }
         // Если NO_OPTION, окно не закрывается
-    }
-
-    private void updateLogContent() {
-        StringBuilder content = new StringBuilder();
-        for (LogEntry entry : m_logSource.all()) {
-            content.append(entry.getMessage()).append("\n");
-        }
-        m_logContent.setText(content.toString());
-        m_logContent.invalidate();
-    }
-
-    @Override
-    public void onLogChanged() {
-        EventQueue.invokeLater(this::updateLogContent);
     }
 }
