+++ src/log/LogChangeListener.java
@@ -2,5 +2,5 @@ package log;
 
 public interface LogChangeListener
 {
-    public void onLogChanged(); 
+    void onLogChanged();
 }
